move encoding registry
