+++ praxis/modules/__init__.py
@@ -1,3 +1,4 @@
-from praxis.modules.common import ENCODING_REGISTRY, MultiIdentity
+from praxis.modules.common import MultiIdentity
 from praxis.modules.embeddings import EMBEDDING_REGISTRY
+from praxis.modules.encoding import ENCODING_REGISTRY
 from praxis.modules.experts import EXPERT_CONFIGS, EXPERT_REGISTRY
